+++ examples/tables/SingleValuesChart.tsx
@@ -72,9 +72,12 @@ const customTheme: ThemeSpec = mergeTheme(downloadTheme, {
         },
     },
     Motion: {
-        mass: 0.5,
-        stiffness: 200,
-        damping: 25,
+        default: {
+            type: 'spring',
+            mass: 0.5,
+            stiffness: 200,
+            damping: 25,
+        },
     },
 })