+++ packages/core/tests/charts/MilestoneMotion.test.tsx
@@ -153,7 +153,7 @@ describe('MilestoneMotion', () => {
                     initial={{ opacity: 0.5, scale: 0.5 }}
                     initialOn={'entry'}
                     visible={false}
-                    transition={{ duration: 0.5, stiffness: 50 }}
+                    transition={{ type: 'spring', duration: 0.5, stiffness: 50 }}
                 >
                     <rect width={10} height={10} />
                 </MilestoneMotion>